+++ CHANGELOG.md
@@ -6,7 +6,35 @@ Release Notes
 **New**
 
 - `Database.indexes(on:)` returns the indexes defined on a database table
+
 - `Database.table(_:hasUniqueKey:)` returns true if a sequence of columns uniquely identifies a row, that is to say if the columns are the primary key, or if there is a unique index on them.
+
+- MutablePersistable types, including Record subclasses, support partial updates:
+    
+    ```swift
+    try person.update(db)                    // Full update
+    try person.update(db, columns: ["age"])  // Only updates the age column
+    ```
+
+**Breaking Changes**
+
+- MutablePersistable `update` and `performUpdate` methods have changed their signature:
+    
+    ```diff
+     protocol MutablePersistable : TableMapping {
+    -func update(db: Database) throws
+    +func update(db: Database, columns: Set<String>) throws
+     }
+     
+     extension MutablePersistable {
+     func update(db: Database) throws
+    +func update(db: Database, columns: Set<String>) throws
+    +func update<S: SequenceType where S.Generator.Element == SQLColumn>(db: Database, columns: S) throws
+    +func update<S: SequenceType where S.Generator.Element == String>(db: Database, columns: S) throws
+    -func performUpdate(db: Database) throws
+    +func performUpdate(db: Database, columns: Set<String>) throws
+     }
+    ```
 
 
 ## 0.76.0