+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 Release Notes
 =============
 
-## Next Version
+## 0.77.0
+
+Released July 28, 2016
 
 **New**
 
-- `Database.indexes(on:)` returns the indexes defined on a database table
+- `Database.indexes(on:)` returns the indexes defined on a database table.
 
 - `Database.table(_:hasUniqueKey:)` returns true if a sequence of columns uniquely identifies a row, that is to say if the columns are the primary key, or if there is a unique index on them.